完成用户管理界面的大部分功能 还剩删除用户功能未完成

+++ creatormc-blog-admin/src/components/main/TopBar.vue
@@ -73,12 +73,14 @@ export default {
      */
     logout() {
       logout().then((response) => {
-        localStorage.removeItem("token");
-        router.push("/login");
-        const store = userStore();
-        //恢复store状态到初始值
-        store.$reset();
-        sessionStorage.clear();
+        if(response != null) {
+          localStorage.removeItem("token");
+          router.push("/login");
+          const store = userStore();
+          //恢复store状态到初始值
+          store.$reset();
+          sessionStorage.clear();
+        }
       });
     }
   },
